Insertion of an element in a linked list
This code signifies how we can insert an element in a linked list at any given valid position. Here we are using two functions mainly the display() function which displays the linked list after the insertions and insertion() function  that takes the data and position as input from the user and inserts the element at it's desired position.